Adding student results page (Teacher's view)
I have added the student results page (teachers view).

I have created the file localy on my computer while there was problems with tailwind css updates. THis page is running good in this branch and on my computer, but thes is some visually issues that came up when i added the code to this branch.