button color on select-user.html
fixing the border color for the page buttons so that they display the correct colors.
color was overidden by a different style. fixed it by putting priority on the border-color in the class.

also changes to hover state on the search for user button as also lacked priority on color.

Same with the "Edit" buttons, they lacked priority on their color aswell.