created page for "add user" button for singular cases.
created the page for creating a user when only needing to create 1, as the pages we had seemed to only add users based on spreadsheet. and since in the design it asked for a "add user" page i went with this. it's basicaly a copy of the manage-user.html just changed things up to fit with this case.

also added href for the button to link the pages between eachother.
so the linking goes like this now:
select-user > create-user = click on "Add user" button.
create-user > select-user = click on either Create or Abort button, both link to the select-user page.