made changes to make it look more like the design
I have added a few things and removed a few lines from the old code.

i have added the nav bar on the left, not linked it to anything, purely styling.
i have fixed some padding and margin to make the spacing between everything as close to pixel perfect as i can see.
as it stands now, the styling for the header section does not seem to be complete. so have to look at it in the future most likely.
i have also gone back and fixed the header section, side nav bar and spacing for the Create user page, so it fits more with the Manage User styling.