User Story 3: Task 1: Implement Contact Form Error Page Structure
Created a Seperate HTML file for displaying error handling on the contact form.

Ensured a structure that follows a logical layout for error handling

Implemented / replaced the submit button with a Try Again button once the form error initiates. This button navigates back to the original contact.html

Replaced header with student header.